+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Unreleased
 - **BREAKING CHANGE** Make define_loader more concise interface like GraphQL's DataLoader.
 - Introduce primary loader through `#define_primary_loader`.
-- Deprecate `#bulk_load_and_compute` in favor of `#bulk_list_and_compute`.
+- **BREAKING CHANGE** Remove `#bulk_load_and_compute` in favor of `#bulk_list_and_compute`.
 
 ## 0.1.1